requirements.txt
absl-py==1.3.0
astunparse==1.6.3
cachetools==5.2.0
cchardet==2.1.7
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.1.1
click==8.1.3
contourpy==1.0.6
cycler==0.10.0
fastprogress==1.0.3
filelock==3.8.0
Flask==2.2.2
flatbuffers==22.10.26
fonttools==4.38.0
gast==0.4.0
glob2==0.7
google-auth==2.14.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.50.0
h5py==3.7.0
huggingface-hub==0.10.1
idna==2.10
importlib-metadata==5.0.0
itsdangerous==2.1.2
jieba==0.42.1
Jinja2==3.1.2
joblib==1.2.0
keras==2.10.0
keras-bert==0.89.0
keras-embed-sim==0.10.0
keras-layer-normalization==0.16.0
keras-multi-head==0.29.0
keras-pos-embd==0.13.0
keras-position-wise-feed-forward==0.8.0
Keras-Preprocessing==1.1.2
keras-self-attention==0.51.0
keras-transformer==0.40.0
kiwisolver==1.4.4
ktrain==0.31.10
langdetect==1.0.9
libclang==14.0.6
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.2
numpy==1.23.4
oauthlib==3.2.2
opencv-python-headless==4.6.0.66
opt-einsum==3.3.0
packaging==21.3
pandas==1.5.1
Pillow==9.3.0
protobuf==3.19.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.2
python-decouple==3.6
python-dotenv==0.21.0
pytz==2022.6
PyYAML==6.0
regex==2022.10.31
requests==2.28.1
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
roboflow==0.2.18
rsa==4.9
sacremoses==0.0.53
scikit-learn==1.1.3
scipy==1.9.3
sentencepiece==0.1.97
six==1.16.0
syntok==1.4.4
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-cpu==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.27.0
termcolor==2.1.0
threadpoolctl==3.1.0
tokenizers==0.13.1
tqdm==4.64.1
transformers==4.17.0
typing-extensions==4.4.0
urllib3==1.26.6
Werkzeug==2.2.2
wget==3.2
Whoosh==2.7.4
wrapt==1.14.1
zipp==3.10.0
gdown
